Separate coordinates context between WoT and HoT
In this commit I separate how the coordinates are used within
Walkontable and Handsontable context. The Walkontable renderer needs to
know about negative coordinates as that module render the whole table.
In The Handsontable context, however only coordinates for cells are
important mostly. As in that context exists cell meta objects which
exists only for cells.

Issue: #6998