Fix rows and columns resizing
There was a problem with getting relative cell position for the TH
elements. I've simplified the fallback logic by retrieving the
coordinates from top (or left for rows) overlay when no other overlay
exists. The previous logic assumes that header element can be part of
the master overlay but it's wrong. Indeed the master overlay has
rendered that elements but they are hidden by CSS (visibility: hidden).

Issue: #7038