Fix rows and columns resizing (#7044)
There was a problem with getting a relative cell position for the TH 
elements. I've simplified the fallback logic by retrieving the coordinates 
from top (or left for rows) overlay when no other overlay exists. The 
previous logic assumes that the header element (passed within 
mouseover event) can be part of the master overlay but it's wrong. 
Indeed the master overlay has rendered those elements but they are
hidden by CSS (visibility: hidden). So it's not possible from UI to 
resize headers through the master overlay.

I've fixed tests that wrongly resize the headers by resizing the top or 
left overlay.

Issue: #7038