Fix unselectable headers when all indexes are hidden
The header selection generally contains cell selection coordinates. In a
case when all rows (or columns) are hidden that visual coordinates are
translated to renderable coordinates that do not exist. Hence no header
highlight is generated. In that case, to make a column (or a row) header
highlight, the row and column index has point to the header (the
negative value).

Issue: #7051