Add ability to highlight headers for hidden indexes
Add ability to highlight headers (gray color) when one of the indexes
(row or column) are hidden and the second index is visible.

To support header highlight for that cases the `getCorner` method of the
Selection class is overwritten. There are two approaches to convince the
header Selection to highlight specific headers. One overwrite the method
which controls coordinates for headers (chosen approach for this commit)
and second rewrite/adjust CellCoords, CellRange, and correct all methods
that deal with numbers to support `null` values - to match changes.

Issue: #7051