Revert headers selection when all indexes are trimmed (#7027)
The PR restores headers' selection when the indexes are trimmed. 
The header selection is necessary for context menu or dropdown menu
actions, which depend on coordinates. Moreover, for better user 
experience, it's better to make headers responsive.

I fix the issue by forwarding the coordinates object without changing
the negative coordinates. Thanks to that change I was able to get rid 
of all workarounds (1, 2, 3) caused by index mapper and made for 
selection.

The changes change the selection module in how it processes the 
coordinates. The CellRange class from now can contain negative 
values, except highlight property, which as the documentation says
points to cells only. That's why the coordinates for highlighting are
normalized to 0 when negative values are passed.

Most of the context menu options were normalized about its
disability state.

Issue: #6998

Co-authored-by: Wojciech Szymański <141330+wszymanski@users.noreply.github.com>