Fix the editor misalign (#7084)
The PR fixes the editor position misalign when it's open. This 
happened when the first row or column was hidden or when there is
a bottom overlay enabled. I fixed this by translating visual 
coordinates into renderable, which are then used in the condition 
which adjusts the editor position.

Issue: #6982

Co-authored-by: Wojciech Szymański <141330+wszymanski@users.noreply.github.com>