Fix header selection using RMB  (#7098)
This PR fixes a bug that made it impossible to select a row or column 
header using RMB. This happens when under the clicked row or column
header there was selected cell.

The PR fixes the bug by removing unnecessary code that was needed
before the negative coordinates feature (#7027).

Additionally, I found that our test helper which emulates mouse events
contains the wrong logic. The helper supports "which" property which
is deprecated. I've replaced to "button" prop.

Issue: #7051