Fix missing overlays in scrollable container
The PR #7075 while fixing double border issues introduced a new bug
which causes missing overlays when the table is scrolled. This happened
only when the window object was set as scrollable element. I reverted
usage of the `getBoundingClientRect` method as this is used to calculate
negative values to detect if the overlay has to be sticked to the client
window edge.

Additionally, this commit fixed bottom overlay which in case when the
window is set as scrollable element the overlay wasn't sticked to the
bottom of the client window. Now it behaves as the top and left overlay.

Issue: #7086