Fix overlays appearance while scrolling and fix bottom overlay positioning  (#7105)
The PR #7075 while fixing double border issues introduced a new bug 
which causes missing overlays when the table is scrolled. This happened
only when the window object was set as a scrollable element. I reverted
usage of the getBoundingClientRect method as this is used to calculate
negative values to detect if the overlay has to be stuck to the client window
edge.

Additionally, this PR fixes the bottom overlay which in the case when the
window is set as a scrollable element the overlay wasn't stuck to the bottom
of the client window. Now it behaves as the top and left overlay.

Issue: #7086