Fix missing bottom border
This commit fixed missing border bottom in a case when the table rows
are trimmed or hidden. In terms of the Walkontable it does not matter as
in both cases total renderable rows are equal to 0. The source of the
problem was the CSS rule controlled by separate overlays which
overwrites the border bottom width alternately.

Issue: #7106