Fix missing bottom border (#7109)
This commit fixes missing overlay bottom border in a case when the table rows
are trimmed or hidden. In terms of the Walkontable, it does not matter as in
both cases total renderable rows are equal to 0. The source of the problem
was the CSS rule controlled by separate overlays that overwrite the 
border-bottom-width alternately. After CSS rule unification, the issue is gone.

Issue: #7106