Fix AutoFill data handling
This commit fixes and issue about the Autofill handler, which in some
cases won't fill the data by dragging the fill handler over the cell.
Additionally, there was some weird behavior with filling the data when
the selection was created from right to left or from bottom to top
direction. After normalization of the selection direction, the AutoFill
plugin starts working as expected.

Issue: #7118