Fix selection while paste into hidden indexes (#7126)
This PR fixes selection desynchronization with data that occurs when the
clipboard was pasted into the table. In a case when in the pasted cell
range, there were hidden indexes, the new selection doesn't cover all
pasted data. I've fixed this by skipping and shifting the hidden indexes by
reading skipColumnsOnPaste and skipRowsOnPaste props. A similar
approach is made in the populateFromArray method of the Core module.

Issue: #7120