Fix infinite loop while scrolling the table
The commit adds additional logic, which checks if the top overlay is
going to an infinite loop caused by added (or removed) `innerBorderTop`
class name. Toggling the class name shifts the viewport by 1px and
triggers the `scroll` event. It causes the table render. The new render
cycle takes into account the shift and toggles the class name again.
This causes the next loops.

When the code detects that the table bottom position is the same as the
overlay bottom position, the loop is interrupted by skipping adjusting
the border-top width through `innerBorderTop` class name.

Issue: #7256