Exclude DOMPurify lib from the handsontable.js build (#7366)
Exclude DOMPurify lib from the dist/handsontable.js build. Adds the lib to the c
opy script. From now on the dompurify while building the package will be copied
to the dist/ directory. And set force to true for the CopyWebpackPlugin webpack
plugin to ensure that dependencies copied from the node_modules are the same as
in the dist directory.