Protect Autocomplete editor from XSS (#7368)
The commit fixes the following issues:

 * Disables values sanitizing when the allowHtml option is enabled (default value
is false). It applies to the Autocomplete and Dropdown editors;
 * Disables values sanitizing for HTML renderers.

Additionally, the commit exports the DOMPurify lib to the window object. This 
will allow adjusting the lib and control of how the HTML content should be 
filtered within the Handsontable settings that accept that kind of strings.

Co-authored-by: Piotr Laszczkowski <swistach@users.noreply.github.com>