Add IndexMap states to the Filters UI components
The commit removes the stateSaver mixin (which wasn't used broadly) and
directly implements the IndexMapper architecture into the Filters UI
components. Moreover, the changes remove the "hiddenRowsCache" map
which, as it turned out, is not necessary at all. The UI components'
hidden state can be calculated on the fly while showing the Dropdown
menu.

Issue: #6830