Prepare es and commonjs dist files
The commit changes how the CommonJS and ES files are transpiled. To
support the new ECMAScript modules, spec ES files from now are
transpiled to files with .mjs extension. The new package.json "exports"
field defines the entry point for the lib.

While preparing the package, I've got a problem with rollup 2, which
won't build the dist files due to the circular dependencies. That's why
I've refactored the Overlay registering functionality. The Overlay
register and constants are moved to separate files.

There is still problem with `hot-formula-parser` module which doesn't
work for webpack4 but it'll be fixed within another commit.

Issue: #7402