Add suspend rendering feature. (#7406)
The commit introduces the concept of the suspend rendering feature. Within the
feature (mentioned in #7274) there is a possibility to narrow the execution of
multiple API calls within one table render cycle.

The feature allows to suspend the following operations:
 * Suspended rendering. The rendering can be suspended using a manual-like and
a framework-like approach.
 * Suspended execution (term execution refers to operations different than DOM
rendering. Currently, only the indexes recalculation allows to postpone the
process).
 * Both.

The suspend rendering feature can reduce the rendering time up to several dozen
times.

* Add rendering and execution batch methods

Issue: #7274

Co-authored-by: Anna Cieślik <scarletfog@gmail.com>