Add temp fix for building dist files
The fix workarounds the bug by manipulating the source code in that way
that the CoreJS polyfill is visible while transpiling the code. Thus to
the import path the file extension is added. In this case when the
"plugins" property was initialized with default keys the polyfill wasn't
visible while files transpiling.

Issue: #7412