Aggregate the plugins enabling
The commit fixes an issue where the error about the missing
dependencies were not thrown in situation when the plugin was not
registered as last.

From now on the Base plugin aggregates the errors from all registered
plugins and then, throws an error if necessary. After that, the plugin
is enabled.

Issue: #7412