The first part of changes to prepare new modularization approach (#7403)
The PR introduces the concept of modularization. This involved the changes in
how the distributed files are built and thus consumed by the developers' app.
As previously, the distributed files are compatible with CommonJS and ES
format. With new changes, they are also compatible with ECMAScript modules
defined in the NodeJS documentation.

The new concept adds the ability to import "base" Handsontable version (`import
Handsontable from 'handsontable/base`;). The Base version, when used, generates
half size smaller bundle than the counterpart full version.

The list of modules that can be imported is as follows:
 * Plugins.;
 * Cell types;
 * Editors;
 * Renderers;
 * Validators;

Each module can be imported using entry point e.q. `import { AutoRowSize,
AutoColumnSize } from 'handsontable/plugins'` or when the developer's
environment doesn't support Tree-Shaking, the modules can be imported
selectively e.q. `import { AutoColumnSize } from
'handsontable/plugins/dropdownMenu'`. The same technique applies to the other
modules.

To make the Plugins possible to use in a modularized environment, the plugin
registration has to be changed. The Plugins, despite the registration order on
the developers' app, are always registered in the same order underneath -
thanks to Priority hardcoded in each plugin.

Co-authored-by: budnix <571316+budnix@users.noreply.github.com>
Co-authored-by: Piotr Laszczkowski <swistach@users.noreply.github.com>