Fix circular dependencies in the i18n module (#7495)
The PR fixes circular dependencies which can be found in the i18n module.

Additionally, the PR fixes the code formatting by splitting the summary exports
into export assigned to particular functions. This prevents jumping between
function declaration or so, and exports.

Moreover, the PR fixes an issue that would show up after the update eslint.
This happens in the monorepo project.