Add custom ESLint rule for restricted module (#7506)
Since the Handsontable was modularized, there are some modules that act as
public API. Importing the main entry of the module can make the "dead" code
elimination process more difficult (especially in older bundlers such as
Webpack3 or in some cases Webpack4). So, to prevent that and to catch developer
mistakes as soon as possible, the custom ESLint rule is added.

I tried using build-in no-restricted-imports rule and rules from the
eslint-plugin-import package but without success. The general problem with that
rules was that it was impossible to restrict import/export modules e.q. from
'../plugins' but at the same time allow import from submodules e.q.
'../plugins/registry'.
I think the most promising rules are in the eslint-plugin-import however while
playing a little bit with that I saw that the error message can not be
customized. Which I think is important for this rule because the "mistake" is
syntactically correct in the context of JavaScript code but in our case, it
generates side-effects that require explanation.

The ESLint doesn't support adding rules or plugins inline into the config file
(the plugin package has to be published). That's why after the NPM install the
plugin is linked. The script is added to the "postinstall" hook.

Additionally, I spotted that the plugins directory was not added to the lining
process. I updated the script and fixes all issues within the files.

Issue: #7473

Co-authored-by: Maciek Łaszcz <nikersify@nikerino.com>