Restructures the Node Modifiers and Index Mapper modules
The commit restructures the Node Modifiers to single files. Thanks to that, the
new modifiers can be easily added without burdening the other modifiers.

The Index Mapper architecture is polished. The Index Maps and Collection Maps
are better categorized. From, now on there is a possibility to create a new
Index Map from the Core API (see `createAndRegisterRowIndexMap` and
`createAndRegisterColumnIndexMap`). Thanks to this, the E2E tests do not have
to use external plugins (e.g., HiddenRows) to simulate hidden, trimmed, or
reorder indexes.

To the IndexMapper, there is added a new method `unregisterAll`. The method
unregisters all Index Maps. It's automatically called when the Handsontable is
destroyed. Thanks to this, the unregistering maps from the all plugins could be
removed.

Issue: #6879