Fix package installation error
The "postinstall" script was not only executed on local environment
while installing necessary dependencies but during the Handsontable
installation on the developer's app side. The script causes an error as
in the package tarball the script file was not included. Moreover the
script file uses native ES syntax so it would cause another problem with
older NodeJS versions.

The fix removes the "postinstall" script and move necessary
swap-package-links.mjs call to user-defined scripts (to those that are
not called automatically by the NPM).