Fix package installation error (#7618)
The postinstall script was not only executed on the local environment while
installing necessary dependencies but during the Handsontable installation on
the developer's app side. This causes an error as in the package tarball the
script file is not included. Moreover, the script file uses native ES syntax so
it would cause another problem with older NodeJS versions.

The fix removes the postinstall script and moves necessary
swap-package-links.mjs call to user-defined scripts (to those that are not
called automatically by the NPM).

Additionally, the commit fixes an issue with the npm run publish script. The 
script cannot be named publish as it's used by the NPM internal live cycle