Add .npmrc config file
The file allows control how the NPM scripts behaves on this repo. The
commit set the "audit-level" setting to "low". Thanks to that the "npm
audit" once detects the vulnerabilities in the deps will exit the
process with an error. The error then is handled by Github Action NPM
Audit script.

Issue: #7621