Add .npmrc config file (#7622)
The file allows control of how the NPM scripts behave on this repo. The
commit set the "audit-level" setting to "low". Thanks to that the "npm
audit" once detects the vulnerabilities in the deps will exit the
process with an error. The error then is handled by Github Action NPM
Audit script.

Issue: #7621