Finish implementing the ChangesObservable module
The module provides the ability to observe the changes that happened in
the index map values during the code running. Currently, only the
"hiding" index map types can be observable. The commit covers all module
implementation with unit tests.

The "crossHidingColumns" meta property is moved from the tree data to
the source settings. Thanks to this, building a tree structure from
scratch (it happens after the "updateSettings" call) does not cause
metadata loss.

To the "hideColumn" and "showColumn" tree modifiers, there is added a
logic that ignores the action when the collapsible node is detected.
This allows postponing the bug fix related to collapsible headers to the
next phase.

Issue: #6879