Add PoC of the changes to the Selection module
That changes fixes the headers selection when the NestedHeaders plugin
is enabled. When the plugin is enabled it blocks the native columns
selection within the Selection module. The plugin is responsible for
column selection. Depends on what header level is clicked and how the
tree structure looks like the plugin based on its state selects the
correct columns.

To achieve that the new hook "beforeHighlightingRowHeader" is added. The
hook allows control the visual column index under the TH element will
be modified.

Issue: #6879