Fix double-tap issue (#7824)
The commit fixes the double-tab issue. It turned out that after calling the
preventDefault method on touchend event the tap can be triggered as fast as a
user can. On the latest Safari, it works cause of the recently added CSS
touch-action property (b805b23) but for other browsers that use the WebView
component, the preventDefault has to be used.

Issue: #7659