Allow to hide columns that have nested headers (#7528)
The commit fixes an issue where hiding columns causes nested headers to glitch.
Several modules had to be extended with new functionalities to get the state
where nested headers work correctly.

 * New modifiers are added to the nodeModifier module. The nested headers tree
structure can be manipulated by the new "hideColumn" and "showColumn"
modifiers. Within the PR, the modifiers are split into separate files for
better maintainability.
 * The "matrixGenerator" module is rewritten to support new tree metadata props
(crossHiddenColumns).
 * For the IndexMapper module, the "changesObservable" module is added. The
module allows observing the index changes changed during HoT runtime.
Currently, only the "hiding" map is supported. Thanks to that the tree does not
have to build from scratch but can be modified on index change.
 * To the IndexMapper, the new methods are added. The methods complement the
IndexMapper API. So from the Handsontable instance, there is a possibility to
create index maps and use them to manipulate rows/columns behavior. For
instance: `hot.rowIndexMapper.createAndRegisterIndexMap`.
 * The header highlighting process is fixed. From, now on only the headers
closest to the dataset are highlighted (grey). An only active highlight (blue)
is applied to all header layers.

It's worth noticing that the PR is focused on fixing the cooperation of the
NestedHeaders plugin and the HidingColumns (and hiding map in general). The
collapsible headers will be fixed in separate PR. Moreover, the PR changes how
the selection of the nested headers is held. Previously no matter what header
level was clicked the selection coordinates started from -1.

The list of backward-compatible changes:
 * New `createAndRegisterIndexMap()` method for IndexMapper;
 * New `createChangesObserver()` method for IndexMapper;
 * New `unregisterAll()` method for IndexMapper;
 * New `beforeHighlightingColumnHeader` and `beforeHighlightingRowHeader` hooks.

Issue: #6879