Fix double-tap issue on iOS (#7980)
The commit adds support for double-tap actions. There is a bug for
mobile browsers that use the WebKit as an engine (excluding the latest
Safari that uses newer version of that engine). The bug causes that a
cell editor can not be open using double-tap events. It's fixed by
calling the "preventDefault" on the "touchend" events. By doing that,
the second tap can be triggered as fast as the user is able to. The
technique removed the restriction where the browsers apply the native
delay ~300ms
(https://webkit.org/blog/5610/more-responsive-tapping-on-ios/).

The "preventDefault" call is not necessary for the newest browsers.
However, it provides a second desired effect. The cells that contain
interactive elements such as "a" link or "button" elements will not be
triggered until the previous highlight cell is the same.

Issue: #7973