Improve the search
The commit splits the search process to two phases each for two
categories. At first, main categories of the API Reference are processed
then if the results limit is not reached, the subcategories are
processed. Thanks to that, we are able to show the suggestions based on
page's importance. The same happens for non-API pages.

From now on, the limit for API and Guides suggestions can be separately
configured.

Issue: #7999