Add support for fuzzy search in the API reference (#8005)
The commit adds the ability to turn on/off the fuzzy search mode for the
specified context. The fuzzy search mode is enabled for API reference for Core,
Hooks, and Options props and methods.
Additionally, the search process is improved. The process is devided into two
phases. At first, the main categories are processed then if the results limit
has not been filled to the limit, the subcategories are processed. Thanks to
that, we are able to show the suggestions based on the page's importance.

Issue: #7999