Improve the Selection demos and fix visual glitches
The commit improves the demo that is used to show how the cells data or
cells meta influences the table. I increase the demo main loop
performance by using the suspend rendering API.

Additionally, the commit includes a fix for demo preview tabs.
Currently, tabs Vue component remembers the last selected tab. When the
remembered tab in other than "Preview", the Handsontable is initialized
within a hidden element. That causes glitches that appear after
switching to the "Preview" tab. The fix internally calls the `render`
method of the Handsontable instance. Thanks to that, the table can be
appropriately calculated.

Issue: #8027