Improve how the related sheets are updated (#8077)
The commit improves how the related sheets (Handsontable instances) are updated. 
I moved the hot update logic from the `onEngineValuesUpdated` hook to separate
hooks that are triggered on data change (dataChange, sourceDataChange). Thanks
to that and the ability to having access to the Handsontable instances
connected to the specific sheets the rendering process could be improved. Thus,
the main issue I worked on #8063 is fixed.

The render method is called only when necessary. For the Handsontable instance
where the data are changed (no matter if it was by API call or UI change), the
call is skipped. When the edited table contains formulas that other sheets
depend on that `hot` instances are rerendered.

Additionally, the PR fixes a bug where the first argument in the
`beforeAutofill` is always a blank array. An original fix that was used to
detect if the fill data was changed caused this bug. To fix this, the most
sensible thing to do is add the argument to the `afterAutofill` hook that
indicates if the fill data is changed pluginHooks.js#L672.

Issue: #8063