Add a separate collection for related sheets for Formula plugin (#8100)
The commit fixes an error that occurs when the HF engine instance was passed to
the Formulas plugin configuration. For that kind of configuration, the
collection that holds the Handsontable instances was not created properly.
Hence, the table that depends on the changed formulas could not be updated.

To solve that issue and with support passing HF instances and Hyperformula
classes in mind, I added a separate collection (based on static registered
module) for connecting HoT instances with the engine. Previously the collection
was mixed with sharedEngineUsage but that collection has a different meaning.

Moreover, I found a memory leak (missing `unregisterEngine` call) that occurs
while calling the `updateSettings` method. I fixed that and cover the changes
with the tests.

Issue: #8097