Fix validation when the Formulas is enabled (#8125)
The PR fixes the validation issues when the Formulas plugin is enabled. My
suggestion to fix that is to restrict the HF engine to manipulate the
Handsontable data only to values that hold formulas. The previous
implementation allows manipulating all cells value by the HF engine. That
approach causes not only the validation problems where the string values like
"2021" were transformed to number 2021 but transforms date-like values such as
"12/1/2016" to 42381 (days after Unix epoch?), and maybe more. This causes
general data desynchronization with what the user passed to the configuration
options.

The PR adds a new method to the public API `isFormulaCellType`.

Moreover, the changes improve the initial execution time by removing
unnecessary cell validation calls (formulas.js#335). The changes are covered
with the tests.

Issue: #8115