Fix all links that points to the old docs (#8236)
The commit fixes all links that point to the old documentation page. Moreover,
the it introduces the following changes:

 * Adds the ability to use the absolute path sign (@) in the markdown links
within the Handsontable source code;
 * Remove the information about cascading configuration from the metaSchema.js
file (I've added referral link to separate page). The "Columns" header in that
description causes scroll misalignment when someone tries to go to the columns
option. As the name was already occupied the true columns option was available
under #columns-2;
 * I added missing content about differences between the dragColumns and
moveColumns methods. The content is copied->pasted from the row-moving.md file
having in mind the methods and hooks name change.