Improve the HoT instances handling (#8286)
The commit improves the code that is responsible for destroying Handsontable
instances used within docs demos. The changes from now on support destroying
not only vanilla demos but supports destroying demos that use wrappers as well.

The registry handler depends on the demo, destroys the Handosntable instance
directly, or for frameworks manually trigger the "destroy" lifecycle. So under
the hood, all framework-related listeners, components, and other resources can
be freed.