Sanity check for destroy after cell validation
We are seeing a rare problem where the instance is somehow destroyed during this `done(valid)` callback. This results in errors like `Cannot read property 'runHooks' of null`.

We can't reproduce in development environment, but we are using the react wrapper and it seems to happen around unmount events. This check should prevent the issue from bubbling up and crashing our app.