Improve the performance of the getCellMeta calls (#8629)
The commit improves the performance of the "getCellMeta" calls by implementing
an additional cache layer. The method is hot and it's called very often from
different parts of the source code. The dynamic nature of the method adds the
ability to extend the cell meta on the fly using the cells function and/or
"beforeGetCellMeta" and "afterGetCellMeta" hooks. That causes performance
bottlenecks.

From my research and observation (I tested the performance using Perf Lab)
adding a simple cache layer that is cleared once per slow render cycle improves
the table performance from ~1% to ~50% (or more). It all depends on how
complicated the logic is implemented within the cells function or if the
"{after/before}GetCellMeta" hooks are used.

Issue: #6303

Co-authored-by: kirszenbaum <jakub.wisniewski@handsontable.com>