Add default font size, color and family (#8681)
The commit changes the default font-size, font-family, font-weight, and color
for all elements within the `.handsontable` CSS class.
Moreover, the PR updates the puppeteer package that uses the latest Chrome
version. Thanks to that I was able to stabilize the tests and get rid of usage
(from some places) of the `toBeArountValue` jasmine matches. The previously
installed version of the puppeteer uses Chrome v80, now it's v93.

Issue: #8661