Filter unwanted files from code docs generation (#8796)
The commit fixes the latest documentation content by creating an allowlist from
the source code that is generated. The latest v10 release introduced several
bugs:

 * Plugins API reference page contains files that do not belong to that
category. Moreover, the listed files are internal modules and should not be
visible at all. Allowlist allows easy control of what and when a specific
module is generated without needing to change the source code, thus without
releasing the package;
* In the search input, there was possible to list internal modules and some
weird files that were created from unknown places of the source code. For
example, type in the search input this phrases "anonymous", "global" or
"module";
* The UndoRedo plugin exposes its methods to the Core but they are not visible
in the API Reference. Now they are included in the Core API ("undo", "redo",
"isUndoAvailable", "isRedoAvailable", "clearUndo");
* The "jsdoc-convert" module contains duplicated files that are removed;
* Improve the script that generates new docs API (npm run docs:api);
* Use shorter GitHub Action docs workflow names with better meaning and ephrase
the names to be sure that all of them will fit in the GitHub Action Workflow tab