Add modularization support for React wrapper (#8854)
The commit modifies the wrapper to use a Base version of the Handsontable
instead of a full-packed version. Thanks to that the wrapper can be customized.
The developer can exclude the modules that are not used within the app thus
decreasing the final size of the vendor chunk file.

The first stage of that feature adds support for providing the modules
globally.