Add `registerAll*` functions to the Base module (#8846)
The commit exposes the new package entry point 'handsontable/registry' that allows registering all build-in modules using only one function call. This change is dictated by the needs of the modularized wrappers, which after modularization will depend on the 'handsontable/base' package, not 'handsontable'. To help developers migrate the new wrappers to the state before modularization there will be possible to register all modules at once instead of registering separately.

The new entry point exposes the following functions:

 * `registerAllEditors` Registers all build-in editors;
 * `registerAllRenderers` Registers all build-in renderers;
 * `registerAllValidators` Registers all build-in validators;
 * `registerAllCellTypes` Registers all build-in cell types;
 * `registerAllPlugins` Registers all build-in plugins;
 * `registerAllModules` Registers all above build-in modules;

Moreover, within this commit, the TS definition files are changed to support a modularized version of Handsontable (#8875). The wrappers from now on are based on the 'handsontable/base' package #8862, #8861, #8854.

Co-authored-by: Kuba Jakub <jakub.wisniewski@handsontable.com>