Fix E2E test watchers (#8990)
The commit fixes an error related to the tests scripts that allow run tests in
watch mode (`npm run test:e2e.watch`). The problem was caused by the little
oldish `trigger-on-stdout-change.js` script. I rewrite the logic of the script
from general usage to more specialized where it re-triggers the `puppeteer`
command after building the test files and source code.

To detect when the webpack finished the build I print non-printable characters
in the standard output that are used to trigger the puppeteer script. The
previous solution was based on the `setTimout` function that obviously was not
precise.

The script uses the already installed dev dependencies. Thanks to that I was
able to remove unused packages `tree-kill`, `support-color`, and
`spawn-command`. Additionally, I replaced the `ecstatic` package - which is
deprecated, with `http-server`.