Fix mobile tests that fail on mobile devices (#9008)
The commit fixes the mobile tests/source code that fails on mobile devices.

According to #8749, the PR fixes most of the tests errors. The exception is
iPad that since releasing the iPad OS on 24th September 2019 the device is no
longer detected as a mobile device (at least by HoT). Thus, it should not be
used to run `./handsontable/test/MobileRunner.html` runner, otherwise, it will
always fail some of the tests.